requirements.txt
langchain
chromadb>=0.4.0
sentence-transformers>=2.2.2
huggingface_hub>=0.11.1
PyMuPDF>=1.22.3
tqdm>=4.65.0
llama-cpp-python>=0.2.74
streamlit